refactor: uriprocessor 패키지 명 변경
