+++ tomcat/src/main/java/org/apache/coyote/http11/urihandler/RootUriHandler.java
@@ -1,6 +1,5 @@
-package org.apache.coyote.http11.uriprocessor;
+package org.apache.coyote.http11.urihandler;
 
-import java.io.IOException;
 import java.util.Map;
 import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.UriResponse;
